fix(submitter): Fix transaction fields (#6306)
### Description

Required fields on EVM transaction are overwritten when we estimate gas
price in EVM adapter for Lander. It means that these fields are not set
when transaction is submitted into a chain.

This change ensures that all the required fields are copied from old
transaction to the new one during gas price estimations.

E2E tests do not pass with fields set, so, we use classical submitter
for the time being.

### Backward compatibility

Yes

### Testing

E2E test

Co-authored-by: Danil Nemirovsky <4614623+ameten@users.noreply.github.com>