fix: token_id as recipient (#6294)
### Description
Currently the Scraper Indexes Warp Transfers incorrectly. It wrongfully
set the `origin_tx_recipient` to the recipient of the warp payment. It
should instead set it to the contract that the TX interacted with -
meaning the `token_id` of the remote transfer.


https://github.com/hyperlane-xyz/hyperlane-monorepo/blob/2faf54edd961f7c738699333f3aefbf765cfb6ef/rust/main/chains/hyperlane-cosmos-native/src/providers/cosmos.rs#L164-L167

The recipient here is for the warp context, meaning the address that
will receive the funds. The contract that the user interacted with is
the `token_id`.
<!--
What's included in this PR?
-->

### Drive-by changes

<!--
Are there any minor or drive-by changes also included?
-->

### Related issues

<!--
- Fixes #[issue number here]
-->

### Backward compatibility

<!--
Are these changes backward compatible? Are there any infrastructure
implications, e.g. changes that would prohibit deploying older commits
using this infra tooling?

Yes/No
-->

### Testing

<!--
What kind of testing have these changes undergone?

None/Manual/Unit Tests
-->