feat: Configure the lowest sequence to index for sequence-aware syncs (#6317)
### Description

When we use Sequence as indexing mode, now we can define the lowest
sequence we would like to index. Before this change, the lowest sequence
was always zero.

It will be the same lowest sequence for all the entities (dispatches,
deliveries and payments). We will re-use field `IndexSettings.from` as
the lowest sequence. Currently, the field is used only for
watermark-aware indexing.

E2E Sealevel configuration was change so that we index sequence 0 since
we start placing transaction from the very beginning.

### Backward compatibility

Yes

### Testing

Unit tests, E2E tests, Manual

---------

Co-authored-by: Danil Nemirovsky <4614623+ameten@users.noreply.github.com>