feat: agent support for `Starknet` chains (#5569)
### Description

- **Starknet Integration**:
- hyperlane-starknet crate has been added, which includes
implementations for Starknet-specific components like mailbox, ISMs ,
MerkleTreeHook, ValidatorAnnounce, and a Starknet provider for
interacting with the chain.
- Indexing and application operation verification logic for Starknet
have been implemented for all agents including the scraper.
- New signer type starkKey has been introduced to handle Starknet
accounts and keys, with corresponding updates in agent key management
(AWS, GCP) and signer configurations. Note: starkKey type keys need to
be generated manually because we lack the tooling for generating it and
storing in KMS right now.

All the changes are now live on testnet for starknetsepolia and
paradexsepolia.


### Drive-by changes

- agent docker tag has been updated for hyperlane validator for
starknetsepolia and paradexsepolia and neutron relayer.

### Related issues

- fixes https://linear.app/hyperlane-xyz/issue/ENG-1219/agents-pr-merge

### Backward compatibility

Yes

### Testing

- e2e test spawn local `starknetdevnet` chains: starknettest23448593 and
starknettest23448594, which send multiple messages to each other.
- e2e tests chain config: RoutingIsm, AggregationIsm, MessageIdMultisig
& MerkleTreeHook
  - e2e tests agents: Relayer, Validator & Scraper

---------

Co-authored-by: 0xevolve <Artevolve@yahoo.com>
Co-authored-by: pbio <10051819+paulbalaji@users.noreply.github.com>
Co-authored-by: yjamin <kaminund@gmail.com>