feat(relayer): config to ignore reported reorgs (#6339)
### Description

There are cases when validators have an unreliable RPC setup and end up
reporting a reorg when no other validator in the set did. We want to
halt processing as a failsafe but also have the option to disable the
failsafe to speed up recovery - hence the config added in this PR, which
applies to a chain when it is the origin of a message.